env: Rmove python-language-server and pygmant
Remove the pygments and python-language-server packages that are not
involved in the project process.

python-language-server has installed for editor completion and
flycheck, pygments has installed to use GNU global for python
parsing. These packages are more relevant to the editor than the
project environment. Therefore, I think it is better to install it in
the editor environment than in the project environment, so I deleted it.